update v1 26/12

+++ services/userCartService.js
@@ -233,9 +233,31 @@ exports.updateItem = async (userId, itemId, qty) => {
         const r = await client.query(q, [itemId, userId]);
         if(r.rows.length == 0) throw Object.assign(new Error('Cart item not found'), { status: 404 });
 
+        //lấy thông tin stock_qty item hiện tại
+        const currentItemRes = await client.query(
+            `SELECT pv.stock_qty AS qty
+            FROM product_variants pv
+            JOIN cart_items ci ON ci.variant_id = pv.id
+            WHERE ci.id = $1 LIMIT 1`,
+            [itemId]
+        );
+        if(currentItemRes.rows.length === 0){
+            throw Object.assign(new Error('Cart item not found'), { status: 404 });
+        }
+        const currentQty = Number(currentItemRes.rows[0].qty);
+
         if(qty === 0 ){
             await client.query(`DELETE FROM cart_items WHERE id = $1`, [itemId]);
-        }else{
+        }//nếu như số lượng mới nhập vào lớn hơn số lượng trong kho thì chỉ cập nhật bằng số lượng trong kho
+        else if (qty > currentQty){
+            await client.query(
+                `UPDATE cart_items
+                SET qty = $1, updated_at = NOW()
+                WHERE id = $2`,
+                [currentQty, itemId]
+            );
+        }
+        else{
             await client.query(
                 `UPDATE cart_items
                 SET qty = $1, updated_at = NOW()
